- stage2: Use ATA driver if any ATA disks are attached. - stage2: Implement and use bios_tty_write() to minimize number of prot-to-real switches.
This fixes AWS T2 boot where bios_disk_read() does not return from int 13h for some reason.
