+++ stage3/Makefile
@@ -4,6 +4,7 @@ PROGRAMS=	stage3.img
 SRCS-stage3.img= \
 	$(CURDIR)/stage3.c \
 	$(OBJDIR)/gitversion.c \
+	$(SRCDIR)/drivers/ata.c \
 	$(SRCDIR)/drivers/ata-pci.c \
 	$(SRCDIR)/drivers/console.c \
 	$(SRCDIR)/drivers/storage.c \
